[GPT-fast benchmark] Merge GPT-fast and micro benchmark output as one csv file
